+++ libs/coin-modules/coin-sui/src/logic/index.ts
@@ -1,4 +1,4 @@
-export { craftTransaction, type CreateExtrinsicArg } from "./craftTransaction";
+export { craftTransaction } from "./craftTransaction";
 export { estimateFees } from "./estimateFees";
 export { broadcast } from "./broadcast";
 export { combine } from "./combine";
@@ -7,3 +7,4 @@ export { lastBlock } from "./lastBlock";
 export { getBlock, getBlockInfo } from "./getBlock";
 export { listOperations } from "./listOperations";
 export { getStakes, getRewards } from "./staking";
+export { canStake } from "./stake";