+++ libs/exchange-module/src/types.ts
@@ -10,10 +10,13 @@ export enum ExchangeType {
 export type ExchangeStartParams =
   | ExchangeStartFundParams
   | ExchangeStartSellParams
-  | ExchangeStartSwapParams;
+  | ExchangeStartSwapParams
+  | ExchangeStartFundParams;
 
 export type ExchangeStartFundParams = {
   exchangeType: "FUND";
+  provider: string;
+  fromAccountId: string;
 };
 
 export type ExchangeStartSellParams = {
@@ -43,6 +46,11 @@ export type ExchangeSwapParams = ExchangeStartSwapParams & {
 };
 
 export type ExchangeStartResult = {
+  transactionId: string;
+  device?: { deviceId?: string; modelId?: string };
+};
+
+export type ExchangeFundResult = {
   transactionId: string;
   device?: { deviceId?: string; modelId?: string };
 };