+++ libs/ledgerjs/packages/hw-app-canton/jest.config.ts
@@ -3,6 +3,7 @@ import baseConfig from "../../jest.config";
 export default {
   ...baseConfig,
   rootDir: __dirname,
+  testPathIgnorePatterns: [...baseConfig.testPathIgnorePatterns, ".*\\.integ\\.test\\.[tj]s"],
   collectCoverageFrom: [
     "src/**/*.ts",
     "!src/**/*.test.{ts,tsx}",