+++ libs/ui/packages/native/src/pre-ldls/components/index.ts
@@ -1,6 +1,7 @@
 export * from "./Input/Input";
 export * from "./AssetItem/AssetItem";
 export * from "./AssetTypeList/AssetList";
+export * from "./ApyIndicator/ApyIndicator";
 export * from "./NetworkItem/NetworkItem";
 export * from "./NetworkList/NetworkList";
 export * from "./Tag/Tag";