+++ libs/coin-tester/CHANGELOG.md
@@ -1,11 +1,18 @@
 # @ledgerhq/coin-tester
 
-## 0.9.1-nightly.0
+## 0.9.1
 
 ### Patch Changes
 
-- Updated dependencies []:
-  - @ledgerhq/hw-transport-node-speculos-http@6.29.10-nightly.0
+- Updated dependencies [[`29d9828`](https://github.com/LedgerHQ/ledger-live/commit/29d9828ff1bde9e7a7171e8e37986ee52f01cd59)]:
+  - @ledgerhq/types-devices@6.27.0
+
+## 0.9.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`29d9828`](https://github.com/LedgerHQ/ledger-live/commit/29d9828ff1bde9e7a7171e8e37986ee52f01cd59)]:
+  - @ledgerhq/types-devices@6.27.0-next.0
 
 ## 0.9.0